+++ packages/react-server-next/package.json
@@ -16,8 +16,8 @@
     },
     "./vite/*": "./dist/vite/*.js",
     "./cloudflare": {
-      "types": "./dist/vite/adapters/cloudflare/runtime.d.ts",
-      "default": "./dist/vite/adapters/cloudflare/runtime.js"
+      "types": "./dist/vite/adapters/cloudflare/platform.d.ts",
+      "default": "./dist/vite/adapters/cloudflare/platform.js"
     },
     ".": {
       "types": "./types/index.d.ts"
