+++ packages/react-server-next/src/vite/adapters/cloudflare/platform-utils.ts
