+++ packages/react-server-next/tsup.config.ts
@@ -7,7 +7,7 @@ export default defineConfig([
       "src/vite/index.ts",
       "src/vite/entry-ssr.tsx",
       "src/vite/adapters/cloudflare/entry.ts",
-      "src/vite/adapters/cloudflare/runtime.ts",
+      "src/vite/adapters/cloudflare/platform.ts",
       "src/vite/adapters/vercel/entry.ts",
       "src/vite/adapters/vercel-edge/entry.ts",
       "src/compat/index.tsx",
